progress handler

+++ src/utils/response.rs
@@ -1,6 +1,6 @@
-use crate::models::response::{Template, TemplateWithData};
+use crate::models::response_model::{Template, TemplateWithData};
 
-pub fn create_response(code: u8, message: &str) -> Template {
+pub fn create_response(code: u16, message: &str) -> Template {
     Template {
         code_http: code,
         message: message.to_string(),
@@ -8,7 +8,7 @@ pub fn create_response(code: u8, message: &str) -> Template {
 }
 
 pub fn create_response_with_data(
-    code: u8,
+    code: u16,
     message: &str,
     data: Option<serde_json::Value>,
 ) -> TemplateWithData {
